removes clients base route and moves overview and sync routes under counts
