+++ ui/app/routes/vault/cluster/clients/counts/sync.ts
@@ -1,0 +1,8 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: BUSL-1.1
+ */
+
+import Route from '@ember/routing/route';
+
+export default class ClientsCountsSyncRoute extends Route {}
