--- ui/app/routes/vault/cluster/clients/overview.js
@@ -1,8 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: BUSL-1.1
- */
-
-import ClientCountBaseRoute from './base';
-
-export default ClientCountBaseRoute;
